requirements.txt

streamlit
pandas
matplotlib
gspread
gspread_dataframe
google-auth
